- move android extensions back to different class - resolve different deprecations
